pyproject.toml
[project]
name = "lark-voice-agent"
version = "0.1.0"
description = "Voice-controlled Lark task management agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.25.0",
    "anthropic>=0.39.0",
    "python-dotenv>=1.0.0",
    "pyobjc-framework-Speech>=10.0",
    "pyobjc-framework-AVFoundation>=10.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
lark-agent = "src.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
